requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
args==0.1.0
async-timeout==5.0.1
attrs==24.3.0
braceexpand==0.1.7
certifi==2024.12.14
charset-normalizer==3.4.1
clint==0.5.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.2.0
-e git+https://github.com/huggingface/diffusers@af28ae2d5ba0ef80d99fff7859ebea730e1cf3f8#egg=diffusers
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
fairscale==0.4.13
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
hpsv2x==1.2.0
huggingface-hub==0.27.1
idna==3.10
-e git+ssh://git@github.com/THUDM/ImageReward.git@2ca71bac4ed86b922fe53ddaec3109fe94d45fd3#egg=image_reward
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
pandas==2.2.1
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
protobuf==3.20.3
pyarrow==18.1.0
pyparsing==3.2.1
pytest==7.2.0
pytest-split==0.8.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sympy==1.13.3
timm==1.0.13
tokenizers==0.15.2
tomli==2.2.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.38.2
triton==3.0.0
tzdata==2024.2
urllib3==2.3.0
webdataset==0.2.100
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
google-genai
typing_extensions